requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pypdf==4.0.1
reportlab==4.0.8
Pillow==10.4.0
gunicorn==21.2.0
PyJWT==2.8.0
bcrypt==4.1.2